fix(web): account TierCard + order page wiring (modifiers, breakdown)
Two more native-parity gaps:

1. Account screen — now renders the same TierCard component used on
   /rewards (port of apps/pickup-native/components/TierCard.tsx),
   matching the native TierCardCarousel that sits below the greeting
   on apps/pickup-native/app/account.tsx. TierCard is hoisted to
   /components/TierCard.tsx so both screens render the same source.
   /rewards/_TierCard.tsx becomes a thin re-export.

   The old custom espresso "Hello" card with inline Beans/Visits was
   too different from native — replaced with a light greeting +
   horizontal Beans/Visits KPIs above the TierCard so the tier
   visual is the hero, like native.

2. Order page — wired the receipt fields that were missing vs
   apps/pickup-native/app/order/[id].tsx: each item now shows its
   modifier selections (joined with " · ") and any specialInstructions
   in italic. Added the discount/SST/total breakdown the native page
   shows: subtotal, reward + reward_name line, tier perk, first-order
   discount, promo discount, SST, total. Surfaces
   loyalty_points_earned beneath the total when the server returned
   one. Order card now uses the same rounded-2xl + 10%-espresso border
   treatment as the product card so the surface family stays
   consistent.

Bumps SW cache to v15.
